Preserve symlinks when extracting from a feature pack.
The new behaviour is to create the symlink as expected. The old behaviour is to
create a file with a single line of text (the single line of text being the
target of the supposed symlink).

I had to change the zip library to commons-compress, as there's no way to tell
if an entry is a symlink or not (no access to the external attribute store)
with the JDK zip library.